Create  crate and extract manifest to it
commit-id:e5ed2bb2
