+++ crates/dojo-cli/src/migrate.rs
@@ -3,8 +3,8 @@ use std::env::{self, current_dir};
 use anyhow::Result;
 use camino::Utf8PathBuf;
 use clap::Args;
-use dojo_project::migration::world::World;
-use dojo_project::{EnvironmentConfig, WorldConfig};
+use dojo_world::migration::world::World;
+use dojo_world::{EnvironmentConfig, WorldConfig};
 use dotenv::dotenv;
 use scarb::core::Config;
 use scarb::ops;
