+++ crates/dojo-lang/Cargo.toml
@@ -24,18 +24,19 @@ cairo-lang-sierra-generator.workspace = true
 cairo-lang-syntax.workspace = true
 cairo-lang-starknet.workspace = true
 cairo-lang-utils.workspace = true
+dojo-world = { path = "../dojo-world" }
 itertools.workspace = true
 scarb.workspace = true
 semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+serde_with = "2.3.1"
 smol_str.workspace = true
 thiserror = "1.0.32"
 tracing.workspace = true
 salsa.workspace = true
 sanitizer = "0.1.6"
 starknet.workspace = true
-serde_with = "2.3.1"
 url = "2.2.2"
 
 [dev-dependencies]
