+++ crates/dojo-lang/src/lib.rs
@@ -7,7 +7,7 @@ mod commands;
 pub mod compiler;
 pub mod component;
 pub mod db;
-pub mod manifest;
+mod manifest;
 pub mod plugin;
 pub mod system;
