+++ crates/dojo-world/Cargo.toml
@@ -1,27 +1,27 @@
 [package]
-name = "dojo-project"
+name = "dojo-world"
 version.workspace = true
 edition.workspace = true
 repository.workspace = true
 license-file.workspace = true
-description = "Dojo project specification. For example, crates and flags used for compilation."
+description = "Dojo world specification. For example, crates and flags used for compilation."
 
 [dependencies]
 anyhow.workspace = true
 cairo-lang-starknet.workspace = true
 cairo-lang-project.workspace = true
 camino.workspace = true
-dojo-lang = { path = "../dojo-lang" }
-serde.workspace = true
-smol_str.workspace = true
 thiserror.workspace = true
 toml = "0.7.1"
 cairo-lang-filesystem.workspace = true
 scarb.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_with = "2.3.1"
+smol_str.workspace = true
 starknet.workspace = true
 url = "2.2.2"
 async-trait = "0.1.68"
-serde_json.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
